Implementing hateos and pagination

+++ src/main/java/com/nasim/repository/ProductRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import com.nasim.model.Product;
 
 
-public interface ProductRepository extends PagingAndSortingRepository<Product, Integer> {
+public interface ProductRepository extends PagingAndSortingRepository<Product, Long> {
 
 }
